requirements.txt
torch>=1.0
torchvision==0.2.2
pretrainedmodels==0.7.4
scikit-learn>=0.20.2
numpy>=1.16
PyYAML==3.13
matplotlib
